src/baseline/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pytorch-lightning>=2.2.0
torchmetrics>=1.3.0
pandas>=2.2.0
pyarrow>=15.0.0
numpy>=1.26.0
Pillow>=10.2.0
opencv-python-headless>=4.9.0.80
albumentations>=1.4.0
tensorboard>=2.16.0
rich>=13.7.0
scikit-learn>=1.4.0
tqdm>=4.66.0
pyyaml>=6.0.1
